feat: add payout, fix event bus publish, bet model

+++ backend/internal/bet/models/bet.go
@@ -4,30 +4,30 @@ import (
 	"fmt"
 )
 
-type BetColor string  
+type BetColor string
 
 const (
-	Red BetColor = "red"
+	Red   BetColor = "red"
 	Black BetColor = "black"
 	Green BetColor = "green"
 )
 
 type BetFill struct {
-	Color BetColor 
-	Number int 
+	Color  BetColor
+	Number int
 }
 type Bet struct {
-	ID int 
-	UserID string 
-	Fill BetFill
-	Stake float64
-	Result bool 
+	ID     int64
+	UserID string
+	Fill   BetFill
+	Stake  float64
+	Result bool
 }
 
 func (b *Bet) Validate() error {
 	if b.ID < 0 {
 		return fmt.Errorf("invalid ID")
-	} 
+	}
 	if b.UserID == "" {
 		return fmt.Errorf("invalid UserID")
 	}
